+++ kir/src/main/java/sacha/kir/bdd/utilisateur/UtilisateurRepository.java
@@ -34,10 +34,10 @@ public interface UtilisateurRepository extends CrudRepository<Utilisateur, Long>
 	@Transactional
 	@Modifying
 	@Query("UPDATE Utilisateur SET Jourscongesrestants = :newstate WHERE UID = :congesid")
-	public void updateJoursConges(@Param("congesid") long uid,@Param("newstate") long joursRestantsConges);
+	public void updateJoursConges(@Param("congesid") long uid,@Param("newstate") float joursRestantsConges);
 	
 	@Transactional
 	@Modifying
 	@Query("UPDATE Utilisateur SET rtt = :newstate WHERE UID = :congesid")
-	public void updateJoursRtt(@Param("congesid") long uid,@Param("newstate") long joursRestantsRtt);
+	public void updateJoursRtt(@Param("congesid") long uid,@Param("newstate") float joursRestantsRtt);
 }
